+++ client/src/components/common/Icon/index.style.ts
