+++ client/src/components/common/Input/index.spec.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Input from '~/components/Input';
+import Input from '~/components/common/Input';
 
 describe('<Input />', () => {
   it('should render same with snapshot', () => {
