+++ client/src/components/login/InputHelp/index.spec.tsx
@@ -3,11 +3,11 @@ import styled from 'styled-components';
 import { render } from '@testing-library/react';
 import 'jest-styled-components';
 
-import Divider from '~/components/Divider';
+import InputHelp from '~/components/login/InputHelp';
 
-describe('<Divider />', () => {
+describe('<InputHelp />', () => {
   it('should render same with snapshot', () => {
-    const { container } = render(<Divider />);
+    const { container } = render(<InputHelp>비밀번호가 너무 짧아요</InputHelp>);
     expect(container).toMatchSnapshot();
   });
 });
