+++ client/src/components/modal/AlertModal/index.style.ts
