+++ client/src/components/modal/ConfirmModal/index.style.ts
