+++ client/src/components/my/EmailInput/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Input from '~/components/Input';
+import Input from '~/components/common/Input';
 import { EmailWrapper } from './index.style';
 
 interface Props {
