+++ client/src/pages/SignUp/index.style.ts
