+++ client/src/styles/app.css
@@ -10,7 +10,7 @@
   flex-direction: column;
   width: 100%;
   height: 100vh;
-  min-height:1024px;
+  min-height: 1024px;
   align-items: center;
 }
